Allow name edition in CE
