+++ packages/strapi-admin/admin/src/components/Roles/RoleForm/NameInput.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import SizedInput from '../../SizedInput';
 
-const NameInput = inputProps => <SizedInput disabled name="name" type="text" {...inputProps} />;
+const NameInput = inputProps => <SizedInput name="name" type="text" {...inputProps} />;
 
 export default NameInput;
