--- packages/strapi-admin/admin/src/components/Roles/RoleForm/utils/getNameInputValue.js
@@ -1,5 +1,0 @@
-const getNameInputValue = (values, role) => {
-  return role.name;
-};
-
-export default getNameInputValue;
