--- packages/strapi-admin/admin/src/containers/Roles/EditPage/utils/getInitialValues.js
@@ -1,7 +1,0 @@
-const getInitialValues = role => {
-  return {
-    description: role.description,
-  };
-};
-
-export default getInitialValues;
