+++ packages/strapi-admin/admin/src/containers/Roles/EditPage/utils/schema.js
@@ -1,7 +1,8 @@
 import * as yup from 'yup';
+import { translatedErrors } from 'strapi-helper-plugin';
 
 const schema = yup.object().shape({
-  name: yup.string(),
+  name: yup.string().required(translatedErrors.required),
 });
 
 export default schema;
