--- packages/strapi-admin/ee/admin/components/Roles/RoleForm/NameInput.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import SizedInput from '../../../../../admin/src/components/SizedInput';
-
-const NameInput = inputProps => <SizedInput name="name" type="text" {...inputProps} />;
-
-export default NameInput;
