--- packages/strapi-admin/ee/admin/components/Roles/RoleForm/utils/getNameInputValue.js
@@ -1,5 +1,0 @@
-const getNameInputValue = values => {
-  return values.name;
-};
-
-export default getNameInputValue;
