--- packages/strapi-admin/ee/admin/containers/Roles/EditPage/utils/getInitialValues.js
@@ -1,8 +1,0 @@
-const getInitialValues = role => {
-  return {
-    name: role.name,
-    description: role.description,
-  };
-};
-
-export default getInitialValues;
