--- packages/strapi-admin/ee/admin/containers/Roles/EditPage/utils/schema.js
@@ -1,8 +1,0 @@
-import * as yup from 'yup';
-import { translatedErrors } from 'strapi-helper-plugin';
-
-const schema = yup.object().shape({
-  name: yup.string().required(translatedErrors.required),
-});
-
-export default schema;
